Address review feedback on tests and inline module size (#299)
Split inline predicates and span helpers into dedicated modules so
inline.rs stays under the 400-line limit, consolidate duplicated wrap
and footnote span tests into rstest tables, and preserve fragment access
via inline re-exports.
